+++ examples/no-tailwind-test/CHANGELOG.md
@@ -4,8 +4,17 @@
 
 ### Patch Changes
 
-- @tour-kit/hints@2.1.1
-- @tour-kit/react@2.1.1
+- Updated dependencies [f62631b]
+- Updated dependencies [e70e310]
+- Updated dependencies [9d1cba1]
+- Updated dependencies [978338b]
+- Updated dependencies [1fe330e]
+- Updated dependencies [9d1cba1]
+- Updated dependencies [978338b]
+- Updated dependencies [1a6e295]
+  - @tour-kit/core@3.0.0
+  - @tour-kit/react@3.0.0
+  - @tour-kit/hints@3.0.0
 
 ## 0.0.23