+++ examples/svelte-app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # svelte-tour-kit-demo
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [3c13df3]
+  - @tour-kit/svelte@1.0.0
+
 ## 0.0.1
 
 ### Patch Changes